qamareth-agents/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
langgraph==0.2.28
langchain-anthropic==0.2.4
anthropic==0.34.2
redis==5.0.8
python-multipart==0.0.12
python-dotenv==1.0.1
pydantic==2.8.2
httpx==0.27.2
